+++ service-a/src/main/java/io/x99/model/UserEntity.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 public class UserEntity {
     @Id
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     private String name;
     private String country;
 
@@ -15,17 +15,17 @@ public class UserEntity {
 
     }
 
-    public UserEntity(Integer id, String name, String country) {
+    public UserEntity(Long id, String name, String country) {
         this.id = id;
         this.name = name;
         this.country = country;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
